support setting a maximum number of readers per path (#1286)
